requirements.txt
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.2.2
blinker==1.4
blis==0.4.1
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cymem==2.0.4
decorator==4.4.2
defusedxml==0.6.0
docx2txt==0.8
entrypoints==0.3
gitdb==4.0.5
GitPython==3.1.12
idna==2.10
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.19.4
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pdfminer.six==20200517
pdfplumber==0.5.25
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
preshed==3.0.4
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.7.0
pyarrow==2.0.0
PyAudio==0.2.11
pycparser==2.20
pycryptodome==3.9.9
pydeck==0.5.0
Pygments==2.7.3
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.5
pyzmq==20.0.0
regex==2020.11.13
requests==2.25.1
scipy==1.5.4
Send2Trash==1.5.0
six==1.15.0
smmap==3.0.4
sortedcontainers==2.3.0
spacy==2.3.2
spicy==0.16.0
srsly==1.0.4
streamlit==0.74.1
streamlit-agraph==0.0.35
streamlit-theme==0.58.0
terminado==0.9.2
testpath==0.4.4
thinc==7.4.1
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.52.0
traitlets==5.0.5
tzlocal==2.1
urllib3==1.26.2
validators==0.18.2
Wand==0.6.5
wasabi==0.8.0
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
